Fixed uid handling Todo now uses uids for todo entries

+++ qml/config.js
@@ -25,10 +25,10 @@ function initialize() {
                 function(tx) {
                     // Create the settings table if it doesn't already exist
                     // If the table exists, this is skipped
-                    tx.executeSql('CREATE TABLE IF NOT EXISTS notes(uid INTEGER UNIQUE, title TEXT,txt TEXT)');
+                    tx.executeSql('CREATE TABLE IF NOT EXISTS notes(uid LONGVARCHAR UNIQUE, title TEXT,txt TEXT)');
                     // TODO: Add uid for title, todo
-                    tx.executeSql('CREATE TABLE IF NOT EXISTS todos(title TEXT,todo TEXT,status INTEGER)');
-                    tx.executeSql('CREATE UNIQUE INDEX IF NOT EXISTS idx ON todos(title,todo);');
+                    tx.executeSql('CREATE TABLE IF NOT EXISTS todos(title TEXT,todo TEXT,status INTEGER,uid LONGVARCHAR)');
+                    tx.executeSql('CREATE UNIQUE INDEX IF NOT EXISTS idx ON todos(title,uid);');
                 });
 }
 
@@ -42,10 +42,10 @@ function setNote(uid,title,txt) {
         var rs = tx.executeSql('INSERT OR REPLACE INTO notes VALUES (?,?,?);', [uid,title,txt]);
         if (rs.rowsAffected > 0) {
             res = "OK";
-            console.log ("Saved to database");
+            //console.log ("Saved to database");
         } else {
             res = "Error";
-            console.log ("Error saving to database");
+            //console.log ("Error saving to database");
         }
     }
     );
@@ -54,22 +54,23 @@ function setNote(uid,title,txt) {
 }
 
 // This function is used to write todos into the database
-function setTodo(title,todo, status) {
+function setTodo(title,todo, status,uid) {
     // profile: name representing the profile of the soundboard
     // nr: int representing the position in the soundboard
     // path: string representing the path to the audio
     // name: string representing the name of the audio
     var db = getDatabase();
     var res = "";
+    //console.debug("todo:" + todo + " uid:" + uid)
     db.transaction(function(tx) {
-        var rs = tx.executeSql('INSERT OR REPLACE INTO todos VALUES (?,?,?);', [title,todo,status]);
+        var rs = tx.executeSql('INSERT OR REPLACE INTO todos VALUES (?,?,?,?);', [title,todo,status,uid]);
         //console.log(rs.rowsAffected)
         if (rs.rowsAffected > 0) {
             res = "OK";
-            console.log ("Saved to database");
+            //console.log ("Saved to database");
         } else {
             res = "Error";
-            console.log ("Error saving to database");
+            //console.log ("Error saving to database");
         }
     }
     );
@@ -85,7 +86,7 @@ function getNotes() {
         var rs = tx.executeSql('SELECT DISTINCT title,uid FROM notes;');
         for (var i = 0; i < rs.rows.length; i++) {
             root.addNote(rs.rows.item(i).title,rs.rows.item(i).uid)
-            console.debug("get Notes:" + rs.rows.item(i).title + " with id:" + rs.rows.item(i).uid)
+            //console.debug("get Notes:" + rs.rows.item(i).title + " with id:" + rs.rows.item(i).uid)
         }
     })
 }
@@ -108,7 +109,7 @@ function getText(title,uid) {
     var notesText="";
     db.transaction(function(tx) {
         var rs = tx.executeSql('SELECT txt FROM notes WHERE title=? AND uid=?;', [title,uid]);
-        console.debug(title + " " + uid)
+        //console.debug(title + " " + uid)
         if (rs.rows.length > 0) {
             notesText = rs.rows.item(0).txt
         } else {
@@ -125,23 +126,26 @@ function getTodo(title) {
     var db = getDatabase();
     var todoText="";
     var todoStatus="";
+    var todoUid="";
     db.transaction(function(tx) {
-        var rs = tx.executeSql('SELECT todo,status FROM todos WHERE title=?;', [title]);
+        var rs = tx.executeSql('SELECT todo,status,uid FROM todos WHERE title=?;', [title]);
         for (var i = 0; i < rs.rows.length; i++) {
             if (rs.rows.length > 0) {
                 var row = rs.rows.item(i)
                 //console.debug(row['todo'])
                 //console.debug(row['status'])
-                todoPage.addTodo(row['todo'],row['status'])
+                //console.debug(row['uid'])
+                todoPage.addTodo(row['todo'],row['status'],row['uid'])
             } else {
                 todoText = "Unknown"
                 todoStatus = 0
+                todoUid = 0
             }
         }
     })
     // The function returns “Unknown” if the setting was not found in the database
     // For more advanced projects, this should probably be handled through error codes
-    return [todoText,todoStatus];
+    return [todoText,todoStatus,todoUid];
 }
 
 // This function is used to remove a note or todo from the database
@@ -149,7 +153,7 @@ function remove(title,type,uid) {
     var db = getDatabase();
     var respath="";
     if (type === "note") {
-        console.debug("Removing Note: " + uid)
+        //console.debug("Removing Note: " + uid)
         db.transaction(function(tx) {
             var rs = tx.executeSql('DELETE FROM notes WHERE title=? AND uid=?;' , [title,uid]);
         })
@@ -162,10 +166,10 @@ function remove(title,type,uid) {
 }
 
 // This function is used to remove todo entry from a todo in the database
-function removeTodoEntry(title,todo) {
+function removeTodoEntry(title,todo,uid) {
     var db = getDatabase();
     var respath="";
     db.transaction(function(tx) {
-        var rs = tx.executeSql('DELETE FROM todos WHERE title=? AND todo=?;' , [title,todo]);
+        var rs = tx.executeSql('DELETE FROM todos WHERE title=? AND todo=? AND uid=?;' , [title,todo,uid]);
     })
 }
